Split models into "conversion" and actual model
Github dependency needs to be removed from the model as they should be
two different things (one can not use the ORM model without pull github)
